+++ src/app/(main)/activity/[semesterId]/[activityId]/boards/[boardId]/create-post/_components/TriggerButton.tsx
