--- src/components/feature/post/create-post-form/activity-date-dialog/index.ts
@@ -1,1 +1,0 @@
-export { ActivityDateFieldDialog } from './ActivityDateDialog'
