--- src/components/feature/post/create-post-form/activity-date-dialog/trigger-button/index.ts
@@ -1,1 +1,0 @@
-export { DateDialogTriggerButton } from './TriggerButton'
